+++ src/main/kotlin/blackjack/domain/gameResult/rule/BlackJackRule.kt
@@ -1,4 +1,4 @@
-package blackjack.domain.gameResult.state
+package blackjack.domain.gameResult.rule
 
 import blackjack.domain.gameResult.GameResult
 import blackjack.domain.participant.Participant
